Refactoring questionaire_screen.dart
