+++ lib/screens/questionaire_screen/questionaire_screen.dart
@@ -20,16 +20,8 @@ import 'dart:core';
 import 'package:flutter/material.dart';
 import 'package:picos/models/phq4.dart';
 import 'package:picos/models/weekly.dart';
-import 'package:picos/screens/questionaire_screen/pages/blood_pressure.dart';
-import 'package:picos/screens/questionaire_screen/pages/body_and_mind.dart';
-import 'package:picos/screens/questionaire_screen/pages/weight.dart';
-import 'package:picos/screens/questionaire_screen/widgets/cover.dart';
-import 'package:picos/screens/questionaire_screen/widgets/questionaire_page.dart';
-import 'package:picos/screens/questionaire_screen/widgets/radio_select_card.dart';
-import 'package:picos/screens/questionaire_screen/widgets/sleep_quality_card.dart';
-import 'package:picos/screens/questionaire_screen/widgets/text_field_card.dart';
+import 'package:picos/screens/questionaire_screen/questionaire_page_state.dart';
 import 'package:picos/widgets/picos_screen_frame.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../models/daily.dart';
 import '../../util/backend.dart';
@@ -45,121 +37,15 @@ class QuestionaireScreen extends StatefulWidget {
 }
 
 class _QuestionaireScreenState extends State<QuestionaireScreen> {
-  //Static Strings
-  static String? _myEntries;
-  static String? _vitalValues;
-  static String? _activityAndRest;
-  static String? _bodyAndMind;
-  static String? _medicationAndTherapy;
-  static String? _ready;
-  static String? _heartFrequency;
-  static String? _bloodSugar;
-  static String? _possibleWalkDistance;
-  static String? _sleepDuration;
-  static String? _hrs;
-  static String? _sleepQuality7Days;
-  static String? _pain;
-  static String? _lowInterest;
-  static String? _dejection;
-  static String? _nervousness;
-  static String? _worries;
-  static String? _changedMedication;
-  static String? _changedTherapy;
-  static Map<String, dynamic>? _painValues;
-  static Map<String, dynamic>? _medicationAndTherapyValues;
-
-  /// Holds the pages to generate the PageView from.
-  static Map<String, Widget>? _pages;
+  QuestionairePageState? _pages;
   static final PageController _controller = PageController();
-
-  /// Maps the pageViews to according titles.
-  static Map<String, String>? _titleMap;
   static const Duration _controllerDuration = Duration(milliseconds: 300);
   static const Curve _controllerCurve = Curves.ease;
-
-  // Value store for the user
-  int? _selectedBodyWeight;
-  int? _selectedBMI;
-  int? _selectedHeartFrequency;
-  int? _selectedSyst;
-  int? _selectedDias;
-  int? _selectedBloodSugar;
-  int? _selectedWalkDistance;
-  int? _selectedSleepDuration;
-  int? _selectedSleepQuality;
-  int? _selectedPain;
-  int? _selectedQuestionA;
-  int? _selectedQuestionB;
-  int? _selectedQuestionC;
-  int? _selectedQuestionD;
 
   // State
   final List<String> _titles = <String>[];
   String? _title;
   final List<Widget> _pageViews = <Widget>[];
-
-  void _initStrings(BuildContext context) {
-    _myEntries = AppLocalizations.of(context)!.myEntries;
-    _vitalValues = AppLocalizations.of(context)!.vitalValues;
-    _activityAndRest = AppLocalizations.of(context)!.activityAndRest;
-    _bodyAndMind = AppLocalizations.of(context)!.bodyAndMind;
-    _medicationAndTherapy = AppLocalizations.of(context)!.medicationAndTherapy;
-    _heartFrequency = AppLocalizations.of(context)!.heartFrequency;
-    _bloodSugar = AppLocalizations.of(context)!.bloodSugar;
-    _possibleWalkDistance = AppLocalizations.of(context)!.possibleWalkDistance;
-    _sleepDuration = AppLocalizations.of(context)!.sleepDuration;
-    _hrs = AppLocalizations.of(context)!.hrs;
-    _ready = AppLocalizations.of(context)!.questionnaireFinished;
-    _sleepQuality7Days = AppLocalizations.of(context)!.sleepQuality7Days;
-    _pain = AppLocalizations.of(context)!.pain;
-    _lowInterest = AppLocalizations.of(context)!.lowInterest;
-    _dejection = AppLocalizations.of(context)!.dejection;
-    _nervousness = AppLocalizations.of(context)!.nervousness;
-    _worries = AppLocalizations.of(context)!.controlWorries;
-    _changedTherapy = AppLocalizations.of(context)!.changedTherapy;
-    _changedMedication = AppLocalizations.of(context)!.changedMedication;
-    _painValues = <String, dynamic>{
-      '0 ${AppLocalizations.of(context)!.painless}': 0,
-      '1 ${AppLocalizations.of(context)!.veryMild}': 1,
-      '2 ${AppLocalizations.of(context)!.unpleasant}': 2,
-      '3 ${AppLocalizations.of(context)!.tolerable}': 3,
-      '4 ${AppLocalizations.of(context)!.disturbing}': 4,
-      '5 ${AppLocalizations.of(context)!.veryDisturbing}': 5,
-      '6 ${AppLocalizations.of(context)!.severe}': 6,
-      '7 ${AppLocalizations.of(context)!.verySevere}': 7,
-      '8 ${AppLocalizations.of(context)!.veryTerrible}': 8,
-      '9 ${AppLocalizations.of(context)!.agonizingUnbearable}': 9,
-      '10 ${AppLocalizations.of(context)!.strongestImaginable}': 10,
-    };
-    _medicationAndTherapyValues = <String, dynamic>{
-      AppLocalizations.of(context)!.yes: true,
-      AppLocalizations.of(context)!.no: false,
-    };
-  }
-
-  void _initTitles(BuildContext context) {
-    _titleMap = <String, String>{
-      'vitalCover': _myEntries!,
-      'weightPage': _vitalValues!,
-      'hearthPage': _vitalValues!,
-      'bloodPressurePage': _vitalValues!,
-      'bloodSugarPage': _vitalValues!,
-      'activityCover': _myEntries!,
-      'walkPage': _activityAndRest!,
-      'sleepDurationPage': _activityAndRest!,
-      'sleepQualityPage': _activityAndRest!,
-      'bodyCover': _myEntries!,
-      'painPage': _bodyAndMind!,
-      'interestPage': _bodyAndMind!,
-      'dejectionPage': _bodyAndMind!,
-      'nervousnessPage': _bodyAndMind!,
-      'worriesPage': _bodyAndMind!,
-      'medicationCover': _myEntries!,
-      'medicationPage': _medicationAndTherapy!,
-      'therapyPage': _medicationAndTherapy!,
-      'readyCover': _myEntries!,
-    };
-  }
 
   void _previousPage() {
     FocusManager.instance.primaryFocus?.unfocus();
@@ -189,220 +75,54 @@ class _QuestionaireScreenState extends State<QuestionaireScreen> {
     );
   }
 
-  void _initPages(BuildContext context) {
-    _pages = <String, Widget>{
-      'vitalCover': Cover(
-        title: _vitalValues!,
-        image: 'assets/Vitalwerte_neg.png',
-        nextFunction: _nextPage,
-      ),
-      'weightPage': Weight(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedBodyWeight: (String value) {
-          _selectedBodyWeight = int.tryParse(value);
-        },
-        onChangedBmi: (String value) {
-          _selectedBMI = int.tryParse(value);
-        },
-      ),
-      'hearthPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: TextFieldCard(
-          label: _heartFrequency!,
-          hint: 'bpm',
-          onChanged: (String value) {
-            _selectedHeartFrequency = int.tryParse(value);
-          },
-        ),
-      ),
-      'bloodPressurePage': BloodPressure(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedSyst: (String value) {
-          _selectedSyst = int.tryParse(value);
-        },
-        onChangedDias: (String value) {
-          _selectedDias = int.tryParse(value);
-        },
-      ),
-      'bloodSugarPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: TextFieldCard(
-          label: _bloodSugar!,
-          hint: 'mg/dL',
-          onChanged: (String value) {
-            _selectedBloodSugar = int.tryParse(value);
-          },
-        ),
-      ),
-      'activityCover': Cover(
-        title: _activityAndRest!,
-        image: 'assets/Aktivitaet+Ruhe_neg.png',
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-      ),
-      'walkPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: TextFieldCard(
-          label: _possibleWalkDistance!,
-          hint: 'Meter',
-          onChanged: (String value) {
-            _selectedWalkDistance = int.tryParse(value);
-          },
-        ),
-      ),
-      'sleepDurationPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: TextFieldCard(
-          label: _sleepDuration!,
-          hint: _hrs!,
-          onChanged: (String value) {
-            _selectedSleepDuration = int.tryParse(value);
-          },
-        ),
-      ),
-      'sleepQualityPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: SleepQualityCard(
-          callBack: (dynamic value) {
-            _selectedSleepQuality = value;
-          },
-          label: _sleepQuality7Days!,
-        ),
-      ),
-      'bodyCover': Cover(
-        title: _bodyAndMind!,
-        image: 'assets/Koerper+Psyche_neg.png',
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-      ),
-      'painPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: RadioSelectCard(
-          callBack: (dynamic value) {
-            _selectedPain = value as int;
-          },
-          label: _pain!,
-          options: _painValues!,
-        ),
-      ),
-      'interestPage': BodyAndMind(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedInterest: (dynamic value) {_selectedQuestionA = value as int;},
-        questionType: _lowInterest!,
-      ),
-      'dejectionPage': BodyAndMind(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedInterest: (dynamic value) {_selectedQuestionB = value as int;},
-        questionType: _dejection!,
-      ),
-      'nervousnessPage': BodyAndMind(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedInterest: (dynamic value) {_selectedQuestionC = value as int;},
-        questionType: _nervousness!,
-      ),
-      'worriesPage': BodyAndMind(
-        previousPage: _previousPage,
-        nextPage: _nextPage,
-        onChangedInterest: (dynamic value) {_selectedQuestionD = value as int;},
-        questionType: _worries!,
-      ),
-      'medicationCover': Cover(
-        title: _medicationAndTherapy!,
-        image: 'assets/Medikation+Therapie_neg.png',
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-      ),
-      'medicationPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: RadioSelectCard(
-          callBack: (dynamic value) {},
-          label: _changedMedication!,
-          options: _medicationAndTherapyValues!,
-        ),
-      ),
-      'therapyPage': QuestionairePage(
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-        child: RadioSelectCard(
-          callBack: (dynamic value) {},
-          label: _changedTherapy!,
-          options: _medicationAndTherapyValues!,
-        ),
-      ),
-      'readyCover': Cover(
-        title: _ready!,
-        image: 'assets/Fertig_Smiley.png',
-        backFunction: _previousPage,
-        nextFunction: _nextPage,
-      ),
-    };
-  }
-
   @override
   Widget build(BuildContext context) {
     // Class init.
-    if (_myEntries == null) {
-      _initStrings(context);
-      _initTitles(context);
-    }
+    _pages ??= QuestionairePageState(_previousPage, _nextPage, context);
 
     // Instance init.
     if (_pageViews.isEmpty) {
-      _initPages(context);
-
-      _title = _myEntries!;
+      _title = _pages!.titleMap!['vitalCover'];
 
       // Add all required pages to the list.
-      _pageViews.add(_pages!['vitalCover']!);
-      _titles.add(_titleMap!['vitalCover']!);
-      _pageViews.add(_pages!['weightPage']!);
-      _titles.add(_titleMap!['weightPage']!);
-      _pageViews.add(_pages!['hearthPage']!);
-      _titles.add(_titleMap!['hearthPage']!);
-      _pageViews.add(_pages!['bloodPressurePage']!);
-      _titles.add(_titleMap!['bloodPressurePage']!);
-      _pageViews.add(_pages!['bloodSugarPage']!);
-      _titles.add(_titleMap!['bloodSugarPage']!);
-      _pageViews.add(_pages!['activityCover']!);
-      _titles.add(_titleMap!['activityCover']!);
-      _pageViews.add(_pages!['walkPage']!);
-      _titles.add(_titleMap!['walkPage']!);
-      _pageViews.add(_pages!['sleepDurationPage']!);
-      _titles.add(_titleMap!['sleepDurationPage']!);
-      _pageViews.add(_pages!['sleepQualityPage']!);
-      _titles.add(_titleMap!['sleepQualityPage']!);
-      _pageViews.add(_pages!['bodyCover']!);
-      _titles.add(_titleMap!['bodyCover']!);
-      _pageViews.add(_pages!['painPage']!);
-      _titles.add(_titleMap!['painPage']!);
-      _pageViews.add(_pages!['interestPage']!);
-      _titles.add(_titleMap!['interestPage']!);
-      _pageViews.add(_pages!['dejectionPage']!);
-      _titles.add(_titleMap!['dejectionPage']!);
-      _pageViews.add(_pages!['nervousnessPage']!);
-      _titles.add(_titleMap!['nervousnessPage']!);
-      _pageViews.add(_pages!['worriesPage']!);
-      _titles.add(_titleMap!['worriesPage']!);
-      _pageViews.add(_pages!['medicationCover']!);
-      _titles.add(_titleMap!['medicationCover']!);
-      _pageViews.add(_pages!['medicationPage']!);
-      _titles.add(_titleMap!['medicationPage']!);
-      _pageViews.add(_pages!['therapyPage']!);
-      _titles.add(_titleMap!['therapyPage']!);
-      _pageViews.add(_pages!['readyCover']!);
-      _titles.add(_titleMap!['readyCover']!);
+      _pageViews.add(_pages!.pages!['vitalCover']!);
+      _titles.add(_pages!.titleMap!['vitalCover']!);
+      _pageViews.add(_pages!.pages!['weightPage']!);
+      _titles.add(_pages!.titleMap!['weightPage']!);
+      _pageViews.add(_pages!.pages!['hearthPage']!);
+      _titles.add(_pages!.titleMap!['hearthPage']!);
+      _pageViews.add(_pages!.pages!['bloodPressurePage']!);
+      _titles.add(_pages!.titleMap!['bloodPressurePage']!);
+      _pageViews.add(_pages!.pages!['bloodSugarPage']!);
+      _titles.add(_pages!.titleMap!['bloodSugarPage']!);
+      _pageViews.add(_pages!.pages!['activityCover']!);
+      _titles.add(_pages!.titleMap!['activityCover']!);
+      _pageViews.add(_pages!.pages!['walkPage']!);
+      _titles.add(_pages!.titleMap!['walkPage']!);
+      _pageViews.add(_pages!.pages!['sleepDurationPage']!);
+      _titles.add(_pages!.titleMap!['sleepDurationPage']!);
+      _pageViews.add(_pages!.pages!['sleepQualityPage']!);
+      _titles.add(_pages!.titleMap!['sleepQualityPage']!);
+      _pageViews.add(_pages!.pages!['bodyCover']!);
+      _titles.add(_pages!.titleMap!['bodyCover']!);
+      _pageViews.add(_pages!.pages!['painPage']!);
+      _titles.add(_pages!.titleMap!['painPage']!);
+      _pageViews.add(_pages!.pages!['interestPage']!);
+      _titles.add(_pages!.titleMap!['interestPage']!);
+      _pageViews.add(_pages!.pages!['dejectionPage']!);
+      _titles.add(_pages!.titleMap!['dejectionPage']!);
+      _pageViews.add(_pages!.pages!['nervousnessPage']!);
+      _titles.add(_pages!.titleMap!['nervousnessPage']!);
+      _pageViews.add(_pages!.pages!['worriesPage']!);
+      _titles.add(_pages!.titleMap!['worriesPage']!);
+      _pageViews.add(_pages!.pages!['medicationCover']!);
+      _titles.add(_pages!.titleMap!['medicationCover']!);
+      _pageViews.add(_pages!.pages!['medicationPage']!);
+      _titles.add(_pages!.titleMap!['medicationPage']!);
+      _pageViews.add(_pages!.pages!['therapyPage']!);
+      _titles.add(_pages!.titleMap!['therapyPage']!);
+      _pageViews.add(_pages!.pages!['readyCover']!);
+      _titles.add(_pages!.titleMap!['readyCover']!);
     }
 
     return PicosScreenFrame(
@@ -423,28 +143,28 @@ class _QuestionaireScreenState extends State<QuestionaireScreen> {
 
             Daily daily = Daily(
               date: date,
-              bloodDiastolic: _selectedDias,
-              bloodSugar: _selectedBloodSugar,
-              bloodSystolic: _selectedSyst,
-              pain: _selectedPain,
-              sleepDuration: _selectedSleepDuration,
-              heartFrequency: _selectedHeartFrequency,
+              bloodDiastolic: _pages!.selectedDias,
+              bloodSugar: _pages!.selectedBloodSugar,
+              bloodSystolic: _pages!.selectedSyst,
+              pain: _pages!.selectedPain,
+              sleepDuration: _pages!.selectedSleepDuration,
+              heartFrequency: _pages!.selectedHeartFrequency,
             );
 
             Weekly weekly = Weekly(
               date: date,
-              bodyWeight: _selectedBodyWeight,
-              bmi: _selectedBMI,
-              sleepQuality: _selectedSleepQuality,
-              walkingDistance: _selectedWalkDistance,
+              bodyWeight: _pages!.selectedBodyWeight,
+              bmi: _pages!.selectedBMI,
+              sleepQuality: _pages!.selectedSleepQuality,
+              walkingDistance: _pages!.selectedWalkDistance,
             );
 
             PHQ4 phq4 = PHQ4(
               date: date,
-              a: _selectedQuestionA,
-              b: _selectedQuestionB,
-              c: _selectedQuestionC,
-              d: _selectedQuestionD,
+              a: _pages!.selectedQuestionA,
+              b: _pages!.selectedQuestionB,
+              c: _pages!.selectedQuestionC,
+              d: _pages!.selectedQuestionD,
             );
 
             await Backend.saveObject(daily);
